% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airports.R
\docType{data}
\name{airports}
\alias{airports}
\title{Airports of the World (from http://www.OpenFlights.org)}
\format{A data frame with 6344 observations on the following 11 variables.}
\source{
\url{http://www.OpenFlights.org/data.html}
}
\usage{
airports
}
\description{
Airports of the World (from http://www.OpenFlights.org)
}
\keyword{datasets}
