% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{apiCheckStatus}
\alias{apiCheckStatus}
\title{Check status of request response.}
\usage{
apiCheckStatus(x)
}
\arguments{
\item{x}{A \code{\link[httr:response]{response()}} object or a number.}
}
\value{
\code{FALSE} if has an error, otherwise \code{TRUE}.
}
\description{
Extract the http status code and convert it into a human readable message. Give
warning if has an error.
}
\examples{
flightscanner:::apiCheckStatus(200L)
flightscanner:::apiCheckStatus(404L)
}
