% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{apiCreateSession}
\alias{apiCreateSession}
\title{Create session of live flight search.}
\usage{
apiCreateSession(origin, destination, startDate, returnDate = NULL,
  adults = 1, children = NULL, infants = NULL, country = "US",
  currency = "USD", locale = "en-US", cabinClass = c("economy",
  "premiumeconomy", "business", "first"), includeCarriers = NULL,
  excludeCarriers = NULL)
}
\arguments{
\item{origin}{(REQUIRED) The origin place, can be country, city, airport, in Skyscanner code.}

\item{destination}{(REQUIRED) The destination, can be country, city, airport, in Skyscanner code.}

\item{startDate}{(REQUIRED) The outbound date. Format \code{"yyyy-mm-dd"}.}

\item{returnDate}{(OPTIONAL) The return date. Format \code{"yyyy-mm-dd"}. Use NULL for oneway
trip.}

\item{adults}{(REQUIRED) Number of adults (16+ years). Must be between 1 and 8.}

\item{children}{(OPTIONAL) Number of children (1-16 years). Can be between 0 and 8.}

\item{infants}{(OPTIONAL) Number of infants (under 12 months). Can be between 0 and 8.}

\item{country}{(REQUIRED) The market/country your user is in.}

\item{currency}{(REQUIRED) The currency you want the prices in.}

\item{locale}{(REQUIRED) The locale you want the results in (ISO locale).}

\item{cabinClass}{(OPTIONAL) The cabin class.
Can be \code{"economy"}, \code{"premiumeconomy"}, \code{"business"}, \code{"first"}.}

\item{includeCarriers}{(OPTIONAL) Only return results from those carriers. Comma-separated list
of carrier ids.}

\item{excludeCarriers}{(OPTIONAL) Filter out results from those carriers. Comma-separated list of
carrier ids.}
}
\value{
A \code{\link[httr:response]{response()}} object of request.
}
\description{
Returns live prices from all our suppliers for the requested flight itinerary. This
function is POST step. Make sure you have set API using \code{\link{apiSetKey}} before.

See \url{https://rapidapi.com/skyscanner/api/skyscanner-flight-search}.
}
\examples{
\dontrun{
apiSetKey("YOUR_API_KEY")
apiCreateSession(origin = "SFO", destination = "LHR", startDate = "2019-07-01")
}
}
\seealso{
\code{\link{apiPollSession}}.
}
