% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{apiMakeHeader}
\alias{apiMakeHeader}
\title{Make headers to a request.}
\usage{
apiMakeHeader(key)
}
\arguments{
\item{key}{API key.}
}
\value{
A character vector contains host, key, and type.
}
\description{
Make headers to a request.
}
\examples{
flightscanner:::apiMakeHeader("YOUR_API_KEY")
}
