% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{apiSet}
\alias{apiSet}
\title{Set API key.}
\usage{
apiSet(key)
}
\arguments{
\item{key}{API key.}
}
\description{
Set API key globally. See
\url{https://rapidapi.com/skyscanner/api/skyscanner-flight-search}.
}
\examples{
\dontrun{
apiSet("YOUR_API_KEY")
}
}
