% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{apiSetKey}
\alias{apiSetKey}
\alias{apiGetKey}
\title{API key.}
\usage{
apiSetKey(key)

apiGetKey()
}
\arguments{
\item{key}{API key.}
}
\value{
\code{apiGetKey} will return the key if it has been set, otherwise \code{NULL}.
}
\description{
Set or get API key in global options. To get the key, see
\url{https://rapidapi.com/skyscanner/api/skyscanner-flight-search}.
}
\examples{
\dontrun{
apiSetKey("YOUR_API_KEY")
apiGetKey()
}
}
