% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{Args2null}
\alias{Args2null}
\title{Convert missing arguments to NULL.}
\usage{
Args2null(x)
}
\arguments{
\item{x}{Argument.}
}
\value{
Converted argument.
}
\description{
Convert missing arguments to NULL.
}
