% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprocess.R
\name{BetweenTime}
\alias{BetweenTime}
\title{Does the time fall in specified range?}
\usage{
BetweenTime(x, interval)
}
\arguments{
\item{x}{A \code{POSIXt} class vector.}

\item{interval}{A character vector of boundary values,
each element should be in format: \code{"\%H:\%M"}.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
This is a compare function of time.
It extracts the hours, minutes and seconds to check if it falls in specified time range.
}
\examples{
x <- lubridate::ymd_hms("2019-01-02 07:30:00")
BetweenTime(x, c("7:30", "8:00"))
BetweenTime(x, c("6:00", "7:29"))
}
