% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprocess.R
\name{CheckDuplicateRow}
\alias{CheckDuplicateRow}
\title{Is there any duplicate rows?}
\usage{
CheckDuplicateRow(.data, .vars)
}
\arguments{
\item{.data}{A data.frame.}

\item{.vars}{Names of Columns to group by. If missing, use the first column name.}
}
\value{
\code{TRUE} if there exists duplicate rows, otherwise \code{FALSE}.
}
\description{
Check if there is any duplicate rows indexed by given columns. If YES, it will give
warning.
}
\examples{
df <- data.frame(Id = c(1, 1), name = c("A", "B"))
flightscanner:::CheckDuplicateRow(df)
flightscanner:::CheckDuplicateRow(df, "name")
}
