% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{CheckStatus}
\alias{CheckStatus}
\title{Check status of request response.}
\usage{
CheckStatus(x)
}
\arguments{
\item{x}{A \code{\link[httr:response]{response()}} object or a number.}
}
\value{
\code{TRUE} if has an error, otherwise \code{FALSE}.
}
\description{
Extract the http status code and convert it into a human readable message.
Give warning if has an error.
}
