% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{CreateJob}
\alias{CreateJob}
\title{Create a cron job.}
\usage{
CreateJob(origin, destination, startDate, returnDate = NULL,
  path = getwd(), frequency = c("daily", "hourly", "minutely"), at,
  ...)
}
\arguments{
\item{origin}{The origin place, can be country, city, airport, in Skyscanner code.}

\item{destination}{The destination, can be country, city, airport, in Skyscanner code.}

\item{startDate}{The outbound date. Format 'yyyy-mm-dd'.}

\item{returnDate}{The return date. Format 'yyyy-mm-dd'. Use NULL for oneway trip.}

\item{path}{Where to put the log, defaults is the current working directory.}

\item{frequency}{A character string equal to one of
\code{"minutely"}, \code{"hourly"}, or \code{"daily"}.}

\item{at}{The actual time of day at which to execute the command.
When unspecified, we default to "8PM",
when the command is to be run less frequently than \code{"hourly"}.}

\item{...}{Other arguments.}
}
\description{
Generate a cron job, and pass it to crontab.
It will automatically scape flight data and save them to a local SQLite database file.
}
\examples{
\dontrun{
CreateJob("SFO", "LHR", "2019-07-01", frequency = "daily", at = "3AM")
CreateJob("SFO", "LHR", "2019-07-01", frequency = "hourly")
CreateJob("SFO", "LHR", "2019-07-01", frequency = "minutely")
}
}
