% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{cron_create}
\alias{cron_create}
\title{Create a cron job.}
\usage{
cron_create(origin, destination, startDate, returnDate = NULL,
  key = apiGetKey(), path = getwd(), frequency = "daily", at, id,
  ...)
}
\arguments{
\item{origin}{The origin place, can be country, city, airport, in Skyscanner code.}

\item{destination}{The destination, can be country, city, airport, in Skyscanner code.}

\item{startDate}{The outbound date. Format 'yyyy-mm-dd'.}

\item{returnDate}{The return date. Format 'yyyy-mm-dd'. Use NULL for oneway trip.}

\item{key}{API key. Default using the key in global options, see \code{\link{apiSetKey}}.}

\item{path}{Where to put the log, defaults is the current working directory.}

\item{frequency}{A character string equal to one of \code{"minutely"}, \code{"hourly"},
\code{"daily"}, \code{"monthly"}, or \code{"yearly"}. Or any complex cron schedule.
.}

\item{at}{The actual time of day at which to execute the command. When unspecified, we default to
\code{"8PM"}, when the command is to be run less frequently than \code{"hourly"}.}

\item{id}{An id, or name, to give to the cronjob task, for easier revision in the future.}

\item{...}{Other arguments.}
}
\description{
Generate a cron job, and pass it to crontab. It will automatically scape flight data
and save them to a local SQLite database file.
}
\examples{
\dontrun{
apiSetKey("YOUR_API_KEY")
cron_create("SFO", "LHR", "2019-07-01", frequency = "daily", at = "3AM")
cron_create("SFO", "LHR", "2019-07-01", frequency = "hourly")
cron_create("SFO", "LHR", "2019-07-01", frequency = "minutely")
cron_create("SFO", "LHR", "2019-07-01", frequency = "0 */2 * * *")  # every 2 hours
}
}
