% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{dbAppendTableNew}
\alias{dbAppendTableNew}
\title{Insert rows into a table.}
\usage{
dbAppendTableNew(conn, name, value, ...)
}
\arguments{
\item{conn}{A \code{\link[DBI:DBIConnection-class]{DBIConnection}} object,
as returned by \code{\link[DBI:dbConnect]{dbConnect()}}.}

\item{name}{Name of the table.}

\item{value}{A data frame of values.
The column names must be consistent with those in the target table in the database.}

\item{...}{Other arguments used by individual methods.}
}
\description{
Insert rows into a table.
If the table has key, it will ignore the duplicated rows indexed by the key.
It assumes that the table has been created beforehand.
}
\examples{
con <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")
dbAppendTableNew(con, "iris", iris)
}
