% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{dbCreateDB}
\alias{dbCreateDB}
\title{Connect to a SQLite database and create tables.}
\usage{
dbCreateDB(conn = RSQLite::SQLite(), dbname = "flight.db")
}
\arguments{
\item{conn}{A \code{\link[RSQLite:SQLiteDriver-class]{SQLiteDriver}} or
\code{\link[RSQLite:SQLiteConnection-class]{SQLiteConnection}}.}

\item{dbname}{The path to the database file. Don't use two exceptions:
\itemize{
  \item{
  \code{""} will create a temporary on-disk database. The file will be deleted when the
  connection is closed.
  }
  \item{
  \code{":memory:"} or \code{"file::memory:"} will create a temporary in-memory database.
  }
}}
}
\value{
An object of class \code{\link[RSQLite:SQLiteConnection-class]{SQLiteConnection}}. If
can't connect to SQLite driver, return 1.
}
\description{
Connect to a SQLite database file. And check it the following tables exists: price,
itinerary, leg, segment, carrier, agent, and place. If not, create them.
}
\examples{
con <- dbCreateDB(dbname = ":memory:")
dbDisconnect(con)
}
