% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{dbSaveData}
\alias{dbSaveData}
\title{Save data to SQLite database.}
\usage{
dbSaveData(conn, x)
}
\arguments{
\item{conn}{A \code{\link[RSQLite:SQLiteConnection-class]{SQLiteConnection}} object, as returned
by \code{\link[DBI:dbConnect]{dbConnect()}}.}

\item{x}{A \code{\link[httr:response]{response()}} object.}
}
\description{
Get data from the request response object, and save data to the corresponding tables
in the SQLite database. Tables includes: price, itinerary, leg, segment, carrier, agent, and
place.
}
\examples{
\dontrun{
# Get response from API
apiSet("YOUR_API_KEY")
resp <- apiCreateSession(origin = "SFO", destination = "LHR", startDate = "2019-07-01")
resp <- apiPollSession(resp)

# Connect to SQLite database
con <- dbCreateDB(dbname = ":memory:")
dbSaveData(con, resp)
dbDisconnect(con)
}
}
