% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download data.}
\usage{
download_data(trip_type, from, to, date)
}
\arguments{
\item{trip_type}{The trip type.}

\item{from}{Where from.}

\item{to}{Where to.}

\item{date}{Date.}
}
\value{
A data frame.
}
\description{
Download data.
}
