% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprocess.R
\name{flightGet}
\alias{flightGet}
\alias{flightGet.response}
\alias{flightGet.SQLiteConnection}
\title{Get flight data from source.}
\usage{
flightGet(x, ...)

\method{flightGet}{response}(x, ...)

\method{flightGet}{SQLiteConnection}(x, ...)
}
\arguments{
\item{x}{An object to get data from.}

\item{...}{Further arguments passed to methods.}
}
\value{
A list of tibbles.
}
\description{
Get flight data from source. Includes: price, itineraries, legs, segments, carriers,
agents, and places. Data are in \code{\link[tibble:tibble]{tibble()}} form.
}
\section{Methods (by class)}{
\itemize{
\item \code{response}: Get flight data from API response.

\item \code{SQLiteConnection}: Get flight data from SQLite connection.
}}

\examples{
\dontrun{
# Get data from API
apiSetKey("YOUR_API_KEY")
resp <- apiCreateSession(origin = "SFO", destination = "LHR", startDate = "2019-07-01")
resp <- apiPollSession(resp)
flightGet(resp)

# Get data from SQLite database
con <- dbCreateDB(dbname = "flight.db")
flightGet(con)
dbDisconnect(con)
}
}
