% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack.R
\name{GetAgents}
\alias{GetAgents}
\title{Extract agent information from request response.}
\usage{
GetAgents(x)
}
\arguments{
\item{x}{A request object.}
}
\value{
A tibble.
}
\description{
Extract information from request response of live flight search endpoint.
Information includes the AgentId, Name, ImageURL, and Type.
}
\examples{
\dontrun{
apiSet("YOUR_API_KEY")
resp <- apiCreateSession(origin = "SFO", destination = "LHR", startDate = "2019-07-01")
resp <- apiPollSession(resp)
flightscanner:::GetAgents(resp)
}
}
\seealso{
\code{\link{GetPrice}}, \code{\link{GetItineraries}}, \code{\link{GetLegs}},
\code{\link{GetSegments}}, \code{\link{GetCarriers}}, \code{\link{GetAgents}},
\code{\link{GetPlaces}}.
}
