% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprocess.R
\name{GetData}
\alias{GetData}
\alias{GetData.response}
\alias{GetData.SQLiteConnection}
\title{Get data from source.}
\usage{
GetData(x, ...)

\method{GetData}{response}(x, ...)

\method{GetData}{SQLiteConnection}(x, lazy = FALSE, ...)
}
\arguments{
\item{x}{An object to get data from.}

\item{...}{Further arguments passed to methods.}

\item{lazy}{If \code{TRUE}, return lazy tibbles.}
}
\value{
A list of tibble.
}
\description{
Get data from source.
Includes: price, itineraries, legs, segments, carriers, agents, and places.
Data are in \code{\link[tibble:tibble]{tibble()}} form.
}
\section{Methods (by class)}{
\itemize{
\item \code{response}: Get data from API response.

\item \code{SQLiteConnection}: Get data from SQLite connection.
}}

\examples{
\dontrun{
# Get data from API
SetAPI("skyscanner-skyscanner-flight-search-v1.p.rapidapi.com", "YOUR_API_KEY")
resp <- CreateSession(origin = "SFO", destination = "LHR", startDate = "2019-07-01")
resp <- PollSession(respondPOST = resp)
GetData(resp)

# Get data from SQLite database
con <- dbCreateDB(dbname = "flight.db")
GetData(con)
}
}
