% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack.R
\name{GetItineraries}
\alias{GetItineraries}
\title{Extract itinerary information from request response.}
\usage{
GetItineraries(x, price = FALSE)
}
\arguments{
\item{x}{A request object.}

\item{price}{If \code{TRUE}, also includes PricingOptions information.}
}
\value{
A tibble.
}
\description{
Extract information from request response of live flight search endpoint.
Information includes the ItineraryId (OutboundLegId, InboundLegId).
}
\examples{
\dontrun{
SetAPI("YOUR_API_KEY")
resp <- apiCreateSession(origin = "SFO", destination = "LHR", startDate = "2019-07-01")
resp <- apiPollSession(resp)
flightscanner:::GetItineraries(resp)
}
}
\seealso{
\code{\link{GetPrice}}, \code{\link{GetItineraries}}, \code{\link{GetLegs}},
\code{\link{GetSegments}}, \code{\link{GetCarriers}}, \code{\link{GetAgents}},
\code{\link{GetPlaces}}.
}
