% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack.R
\name{GetLegs}
\alias{GetLegs}
\title{Extract leg information from request response.}
\usage{
GetLegs(x)
}
\arguments{
\item{x}{A request object.}
}
\value{
A tibble.
}
\description{
Extract information from request response of live flight search endpoint.
Information includes the LegId, SegmentIds, OriginId, DestinationId,
DepartureTime, ArrivalTime, Duration, No.Stops, and Stops.

Stops contains the StopId, and Layover.
}
\examples{
\dontrun{
SetAPI("skyscanner-skyscanner-flight-search-v1.p.rapidapi.com", "YOUR_API_KEY")
resp <- CreateSession(origin = "SFO", destination = "LHR", startDate = "2019-07-01")
resp <- PollSession(respondPOST = resp)
flightscanner:::GetLegs(resp)
}
}
