% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{ListUnpack}
\alias{ListUnpack}
\alias{ListPack}
\title{Write an object to a character string or recreate it from a character string.}
\usage{
ListUnpack(x, mutate = FALSE, vars = NULL)

ListPack(x, mutate = FALSE, vars = NULL,
  vars.time = vars(ends_with("Time")), tz = "UTC")
}
\arguments{
\item{x}{An ojbect or a data.frame.}

\item{mutate}{If \code{TRUE}, apply the function on the given columns. If \code{FALSE}, apply the
function on the object.}

\item{vars, vars.time}{A list of columns generated by \code{\link[dplyr:vars]{vars()}}, a
character vector of column names, a numeric vector of column positions, or \code{NULL}.
\code{vars.time} is the columns names of datetime.}

\item{tz}{a character string that specifies which time zone to parse the date with. The string
must be a time zone that is recognized by the user's OS. If \code{NULL}, use the current time
zone.}
}
\value{
Transformed object.
}
\description{
Writes an ASCII text representation of an R object to a character string, or uses
one to recreate the object. If applying this function to a data.frame, it will do this on each
rows of given columns, and also transform columns between datetime and character.
}
\examples{
s <- flightscanner:::ListUnpack(LETTERS)
flightscanner:::ListPack(s)
d <- flightscanner:::ListUnpack(iris, mutate = TRUE, vars = "Species")
d <- flightscanner:::ListPack(d, mutate = TRUE, vars = "Species")
d$Species <- unlist(d$Species)
head(d)
}
