% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{MakeHeader}
\alias{MakeHeader}
\title{Make headers to a request.}
\usage{
MakeHeader(host, key, type)
}
\arguments{
\item{host}{API hostname.}

\item{key}{API key.}

\item{type}{Content type.}
}
\value{
A character vector contains host, key, and type.
}
\description{
Make headers to a request.
}
