% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{SetAPI}
\alias{SetAPI}
\title{Set API hostname and key.}
\usage{
SetAPI(host, key)
}
\arguments{
\item{host}{API hostname.}

\item{key}{API key.}
}
\value{
A list of hostname and key.
}
\description{
Set API hostname and key globally.

See \url{https://rapidapi.com/skyscanner/api/skyscanner-flight-search}.
}
\examples{
\dontrun{
SetAPI("skyscanner-skyscanner-flight-search-v1.p.rapidapi.com", "YOUR_API_KEY")
}
}
