% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{Args2Null}
\alias{Args2Null}
\title{Convert missing arguments to NULL.}
\usage{
Args2Null(x)
}
\arguments{
\item{x}{Argument.}
}
\value{
Converted argument.
}
\description{
Convert missing arguments to NULL.
}
\examples{
flightscanner:::Args2Null("character")
flightscanner:::Args2Null(NA)
}
