% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprocess.R
\name{CheckDuplicate}
\alias{CheckDuplicate}
\title{Is there any duplicate rows?}
\usage{
CheckDuplicate(.data, .vars)
}
\arguments{
\item{.data}{A data.frame.}

\item{.vars}{Names of Columns to group by. If missing, use the first column name.}
}
\value{
\code{TRUE} if there exists duplicate rows, otherwise \code{FALSE}.
}
\description{
Check if there is any duplicate rows indexed by given columns.
If YES, it will give warning.
}
\examples{
\dontrun{
# Get data from API
SetAPI("skyscanner-skyscanner-flight-search-v1.p.rapidapi.com", "YOUR_API_KEY")
resp <- CreateSession(origin = "SFO", destination = "LHR", startDate = "2019-07-01")
resp <- PollSession(respondPOST = resp)
data <- GetData(resp)
sapply(data, flightscanner:::CheckDuplicate)
}
}
