% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack.R
\name{GetCarriers}
\alias{GetCarriers}
\title{Extract carrier information from request response.}
\usage{
GetCarriers(x)
}
\arguments{
\item{x}{A request object.}
}
\value{
A tibble.
}
\description{
Extract information from request response of live flight search endpoint.
Information includes the CarrierId, Code, Name, and ImageURL.
}
\examples{
\dontrun{
SetAPI("skyscanner-skyscanner-flight-search-v1.p.rapidapi.com", "YOUR_API_KEY")
resp <- CreateSession(origin = "SFO", destination = "LHR", startDate = "2019-07-01")
resp <- PollSession(respondPOST = resp)
flightscanner:::GetCarriers(resp)
}
}
