% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack.R
\name{GetItineraries}
\alias{GetItineraries}
\title{Extract itinerary information from request response.}
\usage{
GetItineraries(x, price = FALSE)
}
\arguments{
\item{x}{A request object.}

\item{price}{If \code{TRUE}, also includes PricingOptions information.}
}
\value{
A tibble.
}
\description{
Extract information from request response of live flight search endpoint.
Information includes the ItineraryId (OutboundLegId, InboundLegId).
}
\examples{
\dontrun{
SetAPI("skyscanner-skyscanner-flight-search-v1.p.rapidapi.com", "YOUR_API_KEY")
resp <- CreateSession(origin = "SFO", destination = "LHR", startDate = "2019-07-01")
resp <- PollSession(respondPOST = resp)
flightscanner:::GetItineraries(resp)
}
}
