% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack.R
\name{GetSegments}
\alias{GetSegments}
\title{Extract segment information from request response.}
\usage{
GetSegments(x)
}
\arguments{
\item{x}{A request object.}
}
\value{
A tibble.
}
\description{
Extract information from request response of live flight search endpoint.
Information includes the SegmentId, OriginId, DestinationId, DepartureTime, ArrivalTime,
Duration, CarrierId, OperatingCarrierId, and FlightNumber.
}
\examples{
\dontrun{
SetAPI("skyscanner-skyscanner-flight-search-v1.p.rapidapi.com", "YOUR_API_KEY")
resp <- CreateSession(origin = "SFO", destination = "LHR", startDate = "2019-07-01")
resp <- PollSession(respondPOST = resp)
flightscanner:::GetSegments(resp)
}
}
