% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{SetAPI}
\alias{SetAPI}
\title{Set API key.}
\usage{
SetAPI(key)
}
\arguments{
\item{key}{API key.}
}
\description{
Set API key globally. See
\url{https://rapidapi.com/skyscanner/api/skyscanner-flight-search}.
}
\examples{
\dontrun{
SetAPI("YOUR_API_KEY")
}
}
